backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence_transformers
numpy
google-generativeai